# docs/executive-intelligence-openapi.yaml
openapi: 3.0.3
info:
  title: Executive Intelligence API
  version: v1.0
  description: |
    Executive Intelligence Platform API Layer — Sprint 6.4C.
    Single entry point for CEO Dashboard, CFO Dashboard, Mobile, AI, Power BI.
    No business logic. Engine-driven. Contract v1.0 FROZEN.
  contact:
    name: Incored ERP Platform Team
    email: julian.pedraza@incored.com.mx

servers:
  - url: https://incored-api.onrender.com/api/executive
    description: Production

security:
  - bearerAuth: []

tags:
  - name: Dashboard
  - name: Intelligence
  - name: Risk
  - name: Portfolio
  - name: Health

paths:
  /health:
    get:
      tags: [Health]
      summary: Engine health check
      description: Returns engine version, providers, capabilities. No financial data.
      security: []
      responses:
        '200':
          description: Health status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /dashboard:
    get:
      tags: [Dashboard]
      summary: Executive dashboard payload
      description: |
        Primary endpoint for CEO/CFO dashboard. Returns ExecutiveDashboardDTO.
        Single call replaces multiple widget requests.
      parameters:
        - $ref: '#/components/parameters/company_id'
        - $ref: '#/components/parameters/fiscal_period'
      responses:
        '200':
          description: ExecutiveDashboardDTO
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/AuthorizationError' }

  /insights:
    get:
      tags: [Intelligence]
      summary: Executive insights
      parameters:
        - $ref: '#/components/parameters/company_id'
        - $ref: '#/components/parameters/fiscal_period'
        - $ref: '#/components/parameters/project_id'
      responses:
        '200':
          description: ExecutiveInsightDTO[] — sorted by priority desc
        '400': { $ref: '#/components/responses/ValidationError' }
        '403': { $ref: '#/components/responses/AuthorizationError' }

  /alerts:
    get:
      tags: [Intelligence]
      summary: Executive alerts
      parameters:
        - $ref: '#/components/parameters/company_id'
        - $ref: '#/components/parameters/fiscal_period'
      responses:
        '200':
          description: ExecutiveAlertDTO[] — OPEN alerts only
        '400': { $ref: '#/components/responses/ValidationError' }

  /rankings:
    get:
      tags: [Intelligence]
      summary: Business entity rankings
      parameters:
        - $ref: '#/components/parameters/company_id'
        - $ref: '#/components/parameters/fiscal_period'
      responses:
        '200':
          description: ExecutiveRankingDTO collections
        '400': { $ref: '#/components/responses/ValidationError' }

  /trends:
    get:
      tags: [Intelligence]
      summary: Financial trend time-series
      parameters:
        - $ref: '#/components/parameters/company_id'
        - $ref: '#/components/parameters/fiscal_period_from'
        - $ref: '#/components/parameters/fiscal_period_to'
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [month, quarter, year]
            default: month
      responses:
        '200':
          description: ExecutiveTrendDTO[]
        '400': { $ref: '#/components/responses/ValidationError' }

  /risk:
    get:
      tags: [Risk]
      summary: Executive risk assessment
      parameters:
        - $ref: '#/components/parameters/company_id'
        - $ref: '#/components/parameters/fiscal_period'
      responses:
        '200':
          description: ExecutiveRiskDTO with risk_level and dimensions

  /portfolio:
    get:
      tags: [Portfolio]
      summary: Portfolio summary — all active projects
      parameters:
        - $ref: '#/components/parameters/company_id'
        - $ref: '#/components/parameters/fiscal_period'
      responses:
        '200':
          description: ExecutivePortfolioDTO[]

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    company_id:
      name: company_id
      in: query
      required: true
      schema: { type: integer, minimum: 1 }
      description: Company scope (validated against authenticated user)
    fiscal_period:
      name: fiscal_period
      in: query
      schema: { type: string, pattern: '^\d{4}-\d{2}$', example: '2026-06' }
    fiscal_period_from:
      name: fiscal_period_from
      in: query
      required: true
      schema: { type: string, pattern: '^\d{4}-\d{2}$' }
    fiscal_period_to:
      name: fiscal_period_to
      in: query
      required: true
      schema: { type: string, pattern: '^\d{4}-\d{2}$' }
    project_id:
      name: project_id
      in: query
      schema: { type: integer, minimum: 1 }

  responses:
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: { code: INVALID_COMPANY_ID, message: company_id must be a positive integer. }
    AuthorizationError:
      description: Company access denied
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  schemas:
    ErrorResponse:
      type: object
      properties:
        success: { type: boolean, example: false }
        error:
          type: object
          properties:
            code: { type: string }
            message: { type: string }
        metadata:
          type: object
          properties:
            request_id: { type: string, format: uuid }
            correlation_id: { type: string, format: uuid }
            generated_at: { type: string, format: date-time }

    HealthResponse:
      type: object
      properties:
        success: { type: boolean }
        data:
          type: object
          properties:
            status: { type: string, enum: [healthy, degraded, down] }
            engine_version: { type: string }
            schema_version: { type: string }
            configuration_provider: { type: string }
            capabilities: { type: array, items: { type: string } }
            timestamp: { type: string, format: date-time }